fix(admin-api): 订单数量放宽为 DECIMAL(10,2)——per_area 小数面积被截断少收
现象：订单明细 quantity 数值语义过窄。`OrderCreateRequest.OrderItemRequest.quantity`
是 Integer、`order_items.quantity` 是 INTEGER，而数量口径按计价方式
（docs/testing/acceptance-protocol.md:225：per_meter=米数 / per_set=1 / per_area=宽×高）。
per_area 的合法面积可以是小数：门幅 2.8m × 3m = 8.4 ㎡，刺绣工艺 30 元/㎡ → 正确
252.00 元；整数列只能表示 8 → 240.00 元 = 少收 12.00 元。

裁定（产品负责人，方案 A）：放宽为 DECIMAL(10,2)/BigDecimal，与 base_price/
subtotal 的既有金额口径一致。

改动：
- DB：V45__widen_order_items_quantity_to_decimal.sql（INTEGER → DECIMAL(10,2)，
  无损扩宽，可在已有数据库安全执行）；schema_full.sql/schema.sql 单一事实源同步。
- Java：OrderCreateRequest/AgentOrderCreateRequest/OrderItem/OrderDetailResponse/
  OrderListResponse 的 quantity → BigDecimal；金额全走 BigDecimal 乘法；
  OrderService.extractProcessingItems() 的 toInteger() 截断改为 toBigDecimal()
  （旧实现把 per_area 的 8.4 截断成 8，列表/详情加工费与外层落库金额自相矛盾），
  toInteger() 已无调用方 → 删除。库存/销量列是整数，仅在写库前取整数部分。
  数量判据由「正整数」改为「大于 0 的数」（向后兼容：JSON 整数 3 仍落 BigDecimal("3")）。
- Agent：order_create 的 items[].quantity / processingItems[].quantity schema
  integer → number；删除「小数数量必须拒绝」闸门；数量口径描述补 per_area=宽×高（㎡）
  与「不得自行取整」；order_query/customer_order_query 的金额兜底不再 int(quantity)。
- 前端：quantity: number（decimal 安全），新建订单页输入本就放行小数，无需改 UI。
- 用例：新增 OR-028（B 端 per_area 小数面积端到端金额回归网）+ 生成物重渲染；
  行为映射 order_create.py 规则补 OR-028（改动必须真跑一次）。
- 契约：CONTRACT-LEDGER 补「订单明细数量」类型契约与「数量类型」行。

测试：
- 新增 OrderQuantityDecimalTest（case_ids: OR-028）——per_area 8.4 落库保真 +
  总额 1092.00（截断成 8 会得 1040.00）、加工项数量解析不截断、
  JSON 整数 3 反序列化兼容、Agent 路径透传、per_meter 2.5 保真。
- Python：per_area 小数保真透传新断言；原「拒绝小数」断言改为「保真不截断」。

+++ backend/admin-api/src/main/java/com/migao/admin/entity/OrderItem.java
@@ -31,7 +31,11 @@ public class OrderItem {
 
     private String productName;
 
-    private Integer quantity;
+    /**
+     * 数量（issue #3666：DECIMAL(10,2)）。口径按计价方式——per_meter=米数、
+     * per_set=1、per_area=宽×高（㎡，可为小数如 8.4）。
+     */
+    private BigDecimal quantity;
 
     private BigDecimal unitPrice;
